✨ feat(Pedidos): Change double
Change double for int to manage all numeric properties in int

+++ RestoStock/Models/Pedidos.cs
@@ -4,7 +4,7 @@
     {
         public int IdProveedor { get; set; }
         public string FechaPedido { get; set; }
-        public double Total { get; set; }
+        public int Total { get; set; }
         public int FkProveedor { get; set; }
         public Proveedores Proveedores { get; set; }
     }
